+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.2.3
 - Add Apple Fruit Decorator to Orchard Trees
+- Make AttachedToLeavesDecorators private
 
 # 2.2.2
 - Remove unneeded BWGPumpkin Class